chore: unify testing code, branch and piece action in same component
