+++ packages/ui/feature-builder-test-steps/src/index.ts
@@ -1,5 +1,4 @@
-export * from './lib/test-piece-step/test-piece-step.component';
-export * from './lib/test-code-step/test-code-step.component';
+export * from './lib/test-action/test-action.component';
 export * from './lib/test-piece-webhook-trigger/test-piece-webhook-trigger.component';
 export * from './lib/ui-feature-builder-test-steps.module';
 export * from './lib/test-steps-core.component';
