--- packages/ui/feature-builder-test-steps/src/lib/test-code-step/test-code-step.component.html
@@ -1,58 +1,0 @@
-<div class="ap-px-4  ap-py-3">
-    <div class="ap-typography-body-1 !ap-font-semibold" i18n>
-        Generate Sample Data
-    </div>
-    <ng-container
-        *ngIf="(testing$ | async) === false  || (testing$ | async) === undefined || (testing$ | async) === null else loadingSpinner">
-        <ng-container
-            *ngIf="(lastTestResult$ |async) !== undefined || (stepTest$ | async)?.standardError; else firstTimeTest">
-            <div class="ap-flex ap-items-center">
-                <div class="ap-min-h-[48px] ap-gap-2  ap-flex ap-items-center">
-                    <ng-container
-                        *ngIf="(stepTest$ | async)?.success === true || (!stepTest$ && (lastTestResult$ | async)) ; else errorTemplate">
-                        <svg-icon src="assets/img/custom/success-step-result.svg"
-                            [svgStyle]="{ width: '21px', height: '20px' }" matTooltip="Step succeeded"></svg-icon>
-                        <span i18n>Tested Successfully</span>
-                    </ng-container>
-                    <ng-template #errorTemplate>
-                        <svg-icon src="assets/img/custom/failure-step-result.svg"
-                            [svgStyle]="{ width: '21px', height: '20px' }" matTooltip="Step failed"></svg-icon>
-                        <span i18n>Test Failed</span>
-                    </ng-template>
-                </div>
-                <div class="ap-flex-grow">
-                </div>
-                <ap-button btnColor="primary" btnStyle="stroked" btnSize="medium"
-                    tooltipText="Test step to generate sample data to use in the next steps"
-                    (buttonClicked)="testStep()" i18n>Retest</ap-button>
-            </div>
-            <div class="ap-typography-body-2 ap-text-body ap--mt-2 ap-mb-4 ap-pl-0.5"
-                *ngIf="!(stepTest$ | async)?.standardError || (stepTest$ | async) === undefined">
-                {{ (lastTestDate$ | async) | timeago:true}}
-            </div>
-            <ap-json-viewer class="json-viewer" title="Output" [maxHeight]="400"
-                [content]="!(stepTest$ | async)?.standardError === true ? (lastTestResult$|async) : (stepTest$ | async)?.standardError"></ap-json-viewer>
-        </ng-container>
-    </ng-container>
-</div>
-
-<ng-template #firstTimeTest>
-    <div class="ap-flex ap-flex-col ap-gap-2 ap-items-center ap-justify-center ap-h-[136px]">
-
-        <div class="ap-relative">
-            <ap-hotspot class="ap-absolute ap-top-[13px] ap-left-[-27px]"
-                matTooltip="Test step to generate sample data to use in the next steps" matTooltipPosition="above">
-            </ap-hotspot>
-            <ap-button btnColor="primary" btnStyle="stroked" btnSize="medium" (buttonClicked)="testStep()" i18n>Test
-                step</ap-button>
-        </div>
-    </div>
-</ng-template>
-<ng-template #loadingSpinner>
-    <div class="ap-flex ap-flex-grow ap-justify-center ap-items-center ap-h-[136px]">
-        <ap-loading-icon class="ap-mr-[-10px]"></ap-loading-icon>
-    </div>
-</ng-template>
-<ng-container *ngIf="stepTest$ | async"></ng-container>
-<ng-container *ngIf="saveStepAfterTesting$ | async"></ng-container>
-<ng-container *ngIf="testDialogClosed$ | async"></ng-container>
