+++ packages/ui/feature-builder-test-steps/src/lib/ui-feature-builder-test-steps.module.ts
@@ -3,16 +3,14 @@ import { CommonModule } from '@angular/common';
 import { UiCommonModule } from '@activepieces/ui/common';
 import { TestPollingTriggerComponent } from './test-polling-trigger/test-polling-trigger.component';
 import { TestWebhookTriggerComponent } from './test-webhook-trigger/test-webhook-trigger.component';
-import { TestPieceStepComponent } from './test-piece-step/test-piece-step.component';
-import { TestCodeStepComponent } from './test-code-step/test-code-step.component';
 import { TestPieceWebhookTriggerComponent } from './test-piece-webhook-trigger/test-piece-webhook-trigger.component';
 import { TimeagoModule } from 'ngx-timeago';
+import { TestActionComponent } from './test-action/test-action.component';
 
 const exportedDeclarations = [
   TestPollingTriggerComponent,
   TestWebhookTriggerComponent,
-  TestPieceStepComponent,
-  TestCodeStepComponent,
+  TestActionComponent,
   TestPieceWebhookTriggerComponent,
 ];
 @NgModule({
